+++ otserv/trunk/spells.cpp
@@ -59,12 +59,12 @@ bool Spells::playerSaySpell(Player* player, SpeakClasses type, const std::string
 {
 	std::string str_words;
 	std::string str_param;
-	unsigned int loc = (uint32_t)words.find( '"', 0 );
+	size_t loc = words.find( '"', 0 );
 	if(loc != std::string::npos && loc >= 0){
 		str_words = std::string(words, 0, loc);
 		str_param = std::string(words, (loc+1), words.size()-loc-1);
 	}
-	else {
+	else{
 		str_words = words;
 		str_param = std::string(""); 
 	}
