+++ otserv/trunk/talkaction.cpp
@@ -92,7 +92,7 @@ TalkActionResult_t TalkActions::playerSaySpell(Player* player, SpeakClasses type
 	
 	std::string str_words;
 	std::string str_param;
-	unsigned int loc = (uint32_t)words.find( '"', 0 );
+	size_t loc = words.find( '"', 0 );
 	if(loc != std::string::npos && loc >= 0){
 		str_words = std::string(words, 0, loc);
 		str_param = std::string(words, (loc+1), words.size()-loc-1);
